Revised and simplied approach on DomainPointer
This change addresses a couple of situations:
a)      where sites had to be created with a freely chosen hostnames, which was
 not possible
b)      we needed a single repository with that would accommodated websites but
also sharepoint and tenant specific OWA sites, which was not possible
c)      The concept of Domain Pointers was ok but appeared hard to understand
for the end users and added some complexity/additional steps when adding a new
site

The domain pointer approach is updated :
A)      decoupled the ?sites? from domain
B)	Created a new ?Domains? Domain Pointer record for each new site created,
 to keep on track of the various sites created through the platform
C)	Upon removal of the site the Pointer gets removed as well
D)	For websites a freely chosen hostname can be taken, similar to
sharepoint (which is or generated or chosen by the tenant)
E)	Removed the pointer displaying/listing from the UI, and show only the
top-level and subdomains.

This changeset contains fixed on the verification if a website or hostheade
(domain pointer) already exists

Fix on if no server component dns record (globaldnsrecord) is defined.

In order to use user selectable hostnames icw dns provisioning use the variable
 [host_name] as hostname in the dns definitions. The variable will be replaced